chore: fix the submenu missplaced and scroll
